Corrected a scanf error in SimpleConnect's tc test program. Corrected a type warning in Connect class.
git-svn-id: https://www.cs.binghamton.edu/svn/simple@884 bb5c8b44-a00f-0410-a998-867a3eb9c5a3

+++ simpleconnect/sc_conn.cpp
@@ -515,7 +515,7 @@ int Connection::RunRecv(void* s)
 		}
 
 		SDL_mutexP(conn->data_mutex);
-		fprintf(stderr,"runnaccept %d, res: %d\n", conn->data.size(), result);
+		fprintf(stderr,"runnaccept %d, res: %d\n", (int)(conn->data.size()), result);
 		map<Uint16, Data>::iterator i = conn->data.begin();
 		for (; i != conn->data.end(); ++i)
 		{
